Homework 3 ioc (#4)
* Add main and di module

* Add annotation

* Add context and ObjectFactory

* Create main

* Update annotation

* Fixed it

* Fixed it

+++ homework-2/sql/02.create-users-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users (
   id BIGSERIAL PRIMARY KEY,
-  credential_id INT UNIQUE,
+  credential_id BIGINT UNIQUE,
   first_name VARCHAR(50),
   last_name VARCHAR(50),
   bio TEXT,
