+++ homework-2/sql/07.create-listing_categories-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE listing_categories (
-  listing_id INT,
-  category_id INT,
+  listing_id BIGINT,
+  category_id BIGINT,
   PRIMARY KEY (listing_id, category_id),
   FOREIGN KEY (listing_id) REFERENCES listings(id),
   FOREIGN KEY (category_id) REFERENCES categories(id)
