+++ homework-2/sql/08.create_comments_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE comments (
   id BIGSERIAL PRIMARY KEY,
-  user_id INT,
-  commenter_id INT,
+  user_id BIGINT,
+  commenter_id BIGINT,
   comment_text TEXT,
   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (user_id) REFERENCES users(id),
