+++ homework-2/sql/09.create-listing_requests-table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE listing_requests (
   id BIGSERIAL PRIMARY KEY,
-  listing_id INT,
-  requester_id INT,
+  listing_id BIGINT,
+  requester_id BIGINT,
   offered_price DECIMAL(10, 2), -- Для покупки и обмена
   status VARCHAR(20) DEFAULT 'pending', -- 'pending', 'accepted', 'rejected'
   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
