+++ homework-2/sql/10.create-transactions-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE transactions (
   id BIGSERIAL PRIMARY KEY,
-  request_id INT,
+  request_id BIGINT,
   completed_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (request_id) REFERENCES listing_requests(id)
 );
