+++ homework-2/sql/13.create-favorite_items-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE favorite_items (
-  user_id INT,
-  listing_id INT,
+  user_id BIGINT,
+  listing_id BIGINT,
   PRIMARY KEY (user_id, listing_id),
   FOREIGN KEY (user_id) REFERENCES users(id),
   FOREIGN KEY (listing_id) REFERENCES listings(id)
